+++ text_annotations/requirements.txt
@@ -10,6 +10,7 @@ dash-table==3.6.0
 decorator==4.4.0
 Flask==1.0.2
 Flask-Compress==1.4.0
+gunicorn==19.9.0
 idna==2.8
 ipython-genutils==0.2.0
 itsdangerous==1.1.0
